# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Meter API",
        "version": "1.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        },
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        },
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        },
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        },
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        },
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        },
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/rechargements/user/{id}": {
            "get": {
                "tags": [
                    "Rechargements"
                ],
                "summary": "Obtenir les rechargements d'un utilisateur",
                "operationId": "8029d215b0db452194ca19986523337a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des rechargements lié a un utilisateur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "schemas/Rechargement"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Utilisateur non trouvé"
                    }
                }
            }
        },
        "/api/rechargements/new": {
            "post": {
                "tags": [
                    "Rechargements"
                ],
                "summary": "Créer un nouveau rechargement",
                "operationId": "52a54e0b23d2d81b53d4bf6c490129a4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "valeur_energ_dispo",
                                    "valeur_energ_acheter",
                                    "montant_recharge",
                                    "date_rechargement",
                                    "heure_rechargement",
                                    "compteur_id",
                                    "user_id"
                                ],
                                "properties": {
                                    "valeur_energ_dispo": {
                                        "type": "number",
                                        "example": 100
                                    },
                                    "valeur_energ_acheter": {
                                        "type": "number",
                                        "example": 50
                                    },
                                    "montant_recharge": {
                                        "type": "number",
                                        "example": 20
                                    },
                                    "date_rechargement": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-01-01"
                                    },
                                    "heure_rechargement": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "12:00:00"
                                    },
                                    "compteur_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rechargement créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "succès"
                                        },
                                        "data": {
                                            "$ref": "schemas/Rechargement"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation échouée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erreur de validation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/releves/compteurs/{id}": {
            "get": {
                "tags": [
                    "Relevés"
                ],
                "summary": "Obtenir les releves d'un compteurs",
                "operationId": "17c144409e1d4528b99d6ce12820e8d2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du compteur",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des releves lié a un compteur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "schemas/Releve"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "compteur non trouvé"
                    }
                }
            }
        },
        "/api/releves": {
            "post": {
                "tags": [
                    "Relevés"
                ],
                "summary": "Créer un nouveau relevé",
                "operationId": "5d4aa86dfaf2d0033b4919f860bc2fa5",
                "requestBody": {
                    "description": "Données requises pour créer un nouveau relevé",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "valeur_energ_dispo",
                                    "date_releve",
                                    "heure_releve",
                                    "compteur_id"
                                ],
                                "properties": {
                                    "valeur_energ_dispo": {
                                        "type": "string",
                                        "example": "100 kWh"
                                    },
                                    "date_releve": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-06-18"
                                    },
                                    "heure_releve": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "13:30:00"
                                    },
                                    "compteur_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Relevé créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "succès"
                                        },
                                        "data": {
                                            "$ref": "schemas/Releve"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Erreur lors de la création du relevé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}